work towards better layout

+++ ReactWindows/ReactNative/UIManager/UIImplementation.cs
@@ -7,6 +7,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Globalization;
 using System.Runtime.CompilerServices;
+using Windows.UI.Xaml.Controls;
 
 namespace ReactNative.UIManager
 {
@@ -109,8 +110,8 @@ namespace ReactNative.UIManager
         /// <param name="eventDispatcher">The event dispatcher.</param>
         public void UpdateRootNodeSize(
             int rootViewTag,
-            int newWidth,
-            int newHeight,
+            float newWidth,
+            float newHeight,
             EventDispatcher eventDispatcher)
         {
             var rootCssNode = _shadowNodeRegistry.GetNode(rootViewTag);
